Create README.md
Hi, my name is Wandersson Sousa Dutra, i'm a Software Engineering student and learning a lot about the process of developing a software and how to make it eficient, easy and achieve it's designed goal. I am also learning c#, a programing lenguage that piqued my interest because of it's robust methods and security. i aim to be a .NET developer who can create various types and multiplatform programs.